requirements.txt
# Core
tensorflow>=2.12,<3.0
numpy>=1.24
pandas>=2.0

# Visualization & reporting
matplotlib>=3.7
seaborn>=0.12
Pillow>=9.0

# Utilities
PyYAML>=6.0
jinja2>=3.1
tqdm>=4.65
tabulate>=0.9

# Optional but useful
# tensorboard typically comes with TensorFlow, but list explicitly for clarity
tensorboard>=2.12
